Void radial density box 5

+++ src/misc/get_subset_patchy.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 list=/hpcstorage/dforero/projects/baosystematics/data/patchy_boxes/used_patchy.dat
-N_mocks=100
-BOX=2
+N_mocks=500
+BOX=1
 mockdir=/hpcstorage/dforero/projects/baosystematics/data/patchy_boxes/PATCHY_CMASS/box${BOX}
 subsetdir=/hpcstorage/dforero/projects/baosystematics/data/patchy_boxes/patchy_cmass_subset/box${BOX}/real
 mkdir -p -v $subsetdir
@@ -9,10 +9,14 @@ echo "Found $(wc -l $list)"
 for ran in $(cat $list | head -${N_mocks})
 do
 filename=$(ls -d $mockdir/* | grep $ran)
-echo $filename
-cp $filename $subsetdir
+#echo $filename
 fn_here=$subsetdir/$(basename $filename)
-bunzip2 $fn_here
+if [[ ! -e $(echo ${fn_here} | sed -e "s/\.bz2//g") ]]
+then
+echo ${fn_here} | sed -e "s/\.bz2//g"
+cp $filename $subsetdir
+bunzip2 -v $fn_here
+fi
 done
 rm -v $subsetdir/*bz2*
